pyproject.toml
[project]
name = "goalscore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "statsmodels>=0.14.6",
]
